Address Gemini review (iteration 3): cleanups and texture-state safety
- PolyLineObjectManager: on the non-opaque path, explicitly disable GL_TEXTURE_2D
  and GL_TEXTURE_COORD_ARRAY so a texture left enabled by a previously drawn manager
  cannot bleed onto the glow lines.
- HorizonLayer: import android.graphics.Color and use the simple Color name.
- HorizonGlowObjectManager: use int locals for index arithmetic, casting to short
  only at the addIndex boundary.
- HorizonGlowPrimitive: rewrap a javadoc line to stay within the 100-column limit.

+++ stardroid-v1/app/src/main/java/com/google/android/stardroid/renderer/HorizonGlowObjectManager.java
@@ -88,21 +88,21 @@ public class HorizonGlowObjectManager extends RendererObjectManager {
 
       // Indices: two triangles per quad between consecutive rings.
       for (int ring = 0; ring < rings.size() - 1; ring++) {
-        short topRowStart = (short) (vertexIndex + ring * ringLength);
-        short bottomRowStart = (short) (topRowStart + ringLength);
+        int topRowStart = vertexIndex + ring * ringLength;
+        int bottomRowStart = topRowStart + ringLength;
         for (int i = 0; i < ringLength - 1; i++) {
-          short topLeft = (short) (topRowStart + i);
-          short topRight = (short) (topLeft + 1);
-          short bottomLeft = (short) (bottomRowStart + i);
-          short bottomRight = (short) (bottomLeft + 1);
+          int topLeft = topRowStart + i;
+          int topRight = topLeft + 1;
+          int bottomLeft = bottomRowStart + i;
+          int bottomRight = bottomLeft + 1;
 
-          indexBuffer.addIndex(topLeft);
-          indexBuffer.addIndex(bottomLeft);
-          indexBuffer.addIndex(bottomRight);
+          indexBuffer.addIndex((short) topLeft);
+          indexBuffer.addIndex((short) bottomLeft);
+          indexBuffer.addIndex((short) bottomRight);
 
-          indexBuffer.addIndex(topLeft);
-          indexBuffer.addIndex(bottomRight);
-          indexBuffer.addIndex(topRight);
+          indexBuffer.addIndex((short) topLeft);
+          indexBuffer.addIndex((short) bottomRight);
+          indexBuffer.addIndex((short) topRight);
         }
       }
       vertexIndex += rings.size() * ringLength;
